Improve LLM JSON parsing and resume heuristics
Add robust JSON extraction for noisy LLM outputs (extract_json_payload) and use it across langchain_client for resume parsing, question generation, evaluation, and report generation. Add heuristic skill extraction (COMMON_TECH_SKILLS + extract_heuristic_skills) in parser_service and merge with structured data. Switch AnswerDetail.score from int to float. Simplify question generator: rely on langchain_client.is_ready for fallback, remove brittle project/company parsing, adjust defaults, and produce resume-specific questions using extracted skills. Update report_service to check rolling LLM readiness. Also add two uploaded resume PDFs. Files changed: langchain_client.py, report.py, generator_service.py, parser_service.py, report_service.py, and two new PDF uploads.

+++ interview-ai-service/app/schemas/report.py
@@ -5,7 +5,7 @@ class AnswerDetail(BaseModel):
     questionText: str = Field(description="The question that was asked")
     answerText: str = Field(description="The answer text transcript provided by the candidate")
     expectedKeywords: List[str] = Field(default=[], description="Expected keywords or technical terms")
-    score: int = Field(description="The evaluated score (0-10) for this answer")
+    score: float = Field(description="The evaluated score (0-10) for this answer")
     feedback: str = Field(description="Constructive feedback for this specific answer")
 
 class ReportGenerationRequest(BaseModel):
